Refactor WheelPicker to add useWheelPicker
